Remove more of the intermediary traces stuff
