+++ frontend/src/pages/RequestDetailsPage/RequestDetailsPageV2/index.tsx
@@ -1,15 +1,9 @@
-import { Hackadoodle } from "./Hackadoodle";
 import { Otel } from "./Otel";
 
 export function RequestDetailsPageV2(props: {
   traceId: string;
-  otel: boolean;
 }) {
-  const { traceId, otel } = props;
+  const { traceId } = props;
 
-  if (otel) {
-    return <Otel traceId={traceId} />;
-  }
-
-  return <Hackadoodle traceId={traceId} />;
+  return <Otel traceId={traceId} />;
 }
