+++ frontend/src/queries/index.ts
@@ -17,7 +17,6 @@ export {
   isMizuFetchLoggingErrorMessage,
   isMizuRequestEndMessage,
   isMizuRequestStartMessage,
-  isMizuLog,
 } from "./types";
 
 export {
@@ -39,15 +38,9 @@ export { useFetchSettings, useUpdateSettings } from "./settings";
 
 export { PROBED_ROUTES_KEY } from "./app-routes";
 export {
-  useMizuTracesV2,
-  type MizuTraceV2,
-  type MizuSpan,
-  type MizuFetchSpan,
   type MizuOrphanLog,
-  isMizuRootRequestSpan,
-  isMizuFetchSpan,
   isMizuOrphanLog,
-} from "./traces-v2";
+} from "./traces-interop";
 
 export {
   type OtelSpan,
